use slowapi for rate limiting instead of raw redis calls. refactor lifespan, middleware, logging
