+++ backend-api/src/logger.py
@@ -1,0 +1,15 @@
+import logging
+import sys
+
+# Configure logging to work with uvicorn
+logger = logging.getLogger("nautichat")
+logger.setLevel(logging.DEBUG)
+
+# Create console handler if it doesn't exist
+if not logger.handlers:
+    console_handler = logging.StreamHandler(sys.stdout)
+    console_handler.setLevel(logging.DEBUG)
+    formatter = logging.Formatter("%(asctime)s [%(levelname)s] %(name)s - %(message)s")
+    console_handler.setFormatter(formatter)
+    logger.addHandler(console_handler)
+    logger.propagate = False  # Prevent double logging in uvicorn
